refator: 자동 저장과 버전 저장 분리 [#11]
