+++ views/layouts/main.hbs
@@ -30,6 +30,7 @@
     <script type="module" src="/javascript/spell/spellCheck.js"></script>
     <script type="module" src="/javascript/spell/popup.js"></script>
     <script type="module" src="/javascript/longSentence/longSentence.js"></script>
-    <script type="module" src="/javascript/storage/storage.js"></script>
+    <script type="module" src="/javascript/storage/autoStorage.js"></script>
+    <script type="module" src="/javascript/storage/versionStorage.js"></script>
 </body>
 </html>
